fix: PacketPreCodec#readShort overflow problem pref: add ServerMultiReceiver and FocessUDPMultiReceiver pref: add interface client to api to give user better access

+++ src/main/java/com/focess/api/net/PacketPreCodec.java
@@ -157,7 +157,7 @@ public class PacketPreCodec {
     public short readShort() {
         short r = 0;
         for (int i = 0; i < 2; i++)
-            r += Byte.toUnsignedInt(data.get(pointer++)) << (i * 8);
+            r += (short) Byte.toUnsignedInt(data.get(pointer++)) << (i * 8);
         return r;
     }
